requirements.txt
shiny==1.4.0
shinywidgets=0.5.2
plotly==6.1.0
shap==0.47.2
pandas==2.2.3
numpy==2.2.5
joblib==1.5.0
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.24
qianfan==0.4.12.3
jinja2==3.1.6